tests: extract strategy_safe_wrapper tests
